debug logging to verify expected behavior matches observed behavior
This is debug code, committed to this branch in case it is useful
someday for debugging purpouses.

The output shows that the C4HudBarsDef is copy constructed to the key
of the deduplication hash.
Because we can't share the same value for the key and the value owned
by the shared pointer, we end up with 2 equal values of the C4HudBarsDef.

Seems this can't be avoided because the key within the unordered_map
is owned by the datastructure and can be moved around by it.

This is why the alive counter increases by 2 for every unique C4HudBarsDef
while it is in use.

The default bars stick around once loaded, until the game exits (as expected).